refactor: Generalized name of the ORM and DB table
