+++ .env.example
@@ -7,3 +7,4 @@ API_KEY=''
 PIPELINE_EXECUTION_INTERVAL_IN_MINUTES=5
 
 AZURE_ODBC_CONNECTION_STRING=''
+AZURE_TABLE_NAME=''
